refactor: domain to sub-packages
