+++ src/main/java/qna/domain/answer/Answer.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.answer;
 
 import java.time.LocalDateTime;
 import javax.persistence.Column;
@@ -8,59 +8,66 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Lob;
 import javax.persistence.Table;
+import qna.domain.question.Question;
+import qna.domain.user.User;
+import qna.exception.NotFoundException;
+import qna.exception.UnAuthorizedException;
 
-@Table(name = "question")
+import java.util.Objects;
+
+@Table(name = "answer")
 @Entity
-public class Question {
+public class Answer {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    @Lob
+    private String contents;
 
     @Column(nullable = false)
     private LocalDateTime createdAt = LocalDateTime.now();
 
     private LocalDateTime updatedAt;
 
-    @Lob
-    private String contents;
-
     @Column(nullable = false)
     private boolean deleted = false;
 
-    @Column(nullable = false, length = 100)
-    private String title;
+    private Long questionId;
 
     private Long writerId;
 
-    protected Question() {
+    public Answer() {
     }
 
-    public Question(String title, String contents, User writer) {
-        this(null, title, contents, writer);
+    public Answer(User writer, Question question, String contents) {
+        this(null, writer, question, contents);
     }
 
-    public Question(Long id, String contents, String title, User writer) {
+    public Answer(Long id, User writer, Question question, String contents) {
+        if (Objects.isNull(writer)) {
+            throw new UnAuthorizedException();
+        }
+        if (Objects.isNull(question)) {
+            throw new NotFoundException();
+        }
         this.id = id;
-        this.contents = contents;
-        this.title = title;
         this.writerId = writer.getId();
+        this.questionId = question.getId();
+        this.contents = contents;
     }
 
-    public boolean isOwner(User writer) {
-        return this.writerId.equals(writer.getId());
+    public boolean isOwner(User loginUser) {
+        return writerId.equals(loginUser.getId());
     }
 
-    public void addAnswer(Answer answer) {
-        answer.toQuestion(this);
+    public void toQuestion(Question question) {
+        this.questionId = question.getId();
     }
 
     public Long getId() {
         return id;
-    }
-
-    public String getTitle() {
-        return title;
     }
 
     public String getContents() {
