+++ src/main/java/qna/domain/deletehistory/ContentType.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.deletehistory;
 
 public enum ContentType {
     QUESTION, ANSWER
