+++ src/main/java/qna/domain/question/QuestionRepository.java
@@ -1,9 +1,10 @@
-package qna.domain;
+package qna.domain.question;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 import java.util.Optional;
+import qna.domain.question.Question;
 
 public interface QuestionRepository extends JpaRepository<Question, Long> {
     List<Question> findByDeletedFalse();
