+++ src/main/java/qna/domain/user/User.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.user;
 
 import java.time.LocalDateTime;
 import javax.persistence.Column;
