+++ src/main/java/qna/service/DeleteHistoryService.java
@@ -3,8 +3,8 @@ package qna.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import qna.domain.DeleteHistory;
-import qna.domain.DeleteHistoryRepository;
+import qna.domain.deletehistory.DeleteHistory;
+import qna.domain.deletehistory.DeleteHistoryRepository;
 
 import java.util.List;
