+++ src/main/java/qna/service/QnaService.java
@@ -4,9 +4,15 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import qna.domain.answer.Answer;
+import qna.domain.answer.AnswerRepository;
+import qna.domain.deletehistory.DeleteHistory;
+import qna.domain.deletehistory.ContentType;
+import qna.domain.question.Question;
+import qna.domain.question.QuestionRepository;
+import qna.domain.user.User;
 import qna.exception.CannotDeleteException;
 import qna.exception.NotFoundException;
-import qna.domain.*;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
