+++ src/test/java/qna/domain/answer/AnswerTest.java
@@ -1,4 +1,7 @@
-package qna.domain;
+package qna.domain.answer;
+
+import qna.domain.question.QuestionTest;
+import qna.domain.user.UserTest;
 
 public class AnswerTest {
     public static final Answer A1 = new Answer(UserTest.JAVAJIGI, QuestionTest.Q1, "Answers Contents1");
