+++ src/test/java/qna/domain/question/QuestionTest.java
@@ -1,4 +1,6 @@
-package qna.domain;
+package qna.domain.question;
+
+import qna.domain.user.UserTest;
 
 public class QuestionTest {
     public static final Question Q1 = new Question("title1", "contents1", UserTest.JAVAJIGI);
